Extracting a Mindgard interface for probes
* This allows for implementing different protocols for probes that is
  supported by Mindgard.
* The hasMoreProbes() is removed as it was only used in tests and not supporting
  the behaviour of the extension.
* The poll timeout is now configurable so that we don't need to expose additional
  methods for testing purposes.
* This changeset also includes some usability improvements around diagnosing any
  problems that might occur in using the extension (showing timestamps on when important
  actions occur)
